### 2021-06-28 #3425 @whedon
[PRE REVIEW]: SiSyPHE: A Python package for the Simulation of Systems of Particles with High Efficiency
**Submitting author:** @antoinediez (<a href="http://orcid.org/0000-0002-4454-8452">Antoine Diez</a>)
**Repository:** <a href="https://github.com/antoinediez/Sisyphe" target ="_blank">https://github.com/antoinediez/Sisyphe</a>
**Version:** v0.1.5
