### 2021-06-28 8838d16
add missing DOIs
