+++ doc/biblio_sisyphe.bib
@@ -66,7 +66,8 @@
 	publisher = {Springer-Verlag Berlin Heidelberg},
 	series = {Lecture {Notes} in {Mathematics}},
 	title = {Asymptotic {Behaviour} of some interacting particle systems; {McKean}-{Vlasov} and {Boltzmann} models},
-	year = {1996}}
+	year = {1996},
+	doi = {10.1007/bfb0093177}}
 
 @article{vicsek_novel_1995,
 	author = {Vicsek, T. and Czir{\'o}k, A. and Ben-Jacob, E. and Cohen, I. and Shochet, O.},
@@ -119,7 +120,8 @@
 	pages = {453--469},
 	title = {Emergence of time-asymptotic flocking in a stochastic {Cucker}-{Smale} system},
 	volume = {7},
-	year = {2009}}
+	year = {2009},
+	doi = {10.4310/cms.2009.v7.n2.a9}}
 
 @incollection{naldi_particle_2010,
 	author = {Carrillo, J. A. and Fornasier, M. and Toscani, G. and Vecil, F.},
@@ -151,7 +153,8 @@
 	pages = {6055--6157},
 	title = {On {Kac}'s chaos and related problems},
 	volume = {266},
-	year = {2014}}
+	year = {2014},
+	doi = {10.1016/j.jfa.2014.02.030}}
 
 @article{degond_phase_2015,
 	author = {Degond, P. and Frouvelle, A. and Liu, J.-G.},
@@ -192,7 +195,8 @@
 	pages = {960--995},
 	title = {{McKean}--{Vlasov} limit for interacting systems with simultaneous jumps},
 	volume = {36},
-	year = {2018}}
+	year = {2018},
+	doi = {10.1080/07362994.2018.1486202}}
 
 @inproceedings{degond_mathematical_2018,
 	address = {Rio de Janeiro, Brazil},
@@ -203,7 +207,8 @@
 	pages = {3943--3964},
 	title = {Mathematical models of collective dynamics and self-organization},
 	volume = {4},
-	year = {2018}}
+	year = {2018},
+	doi = {10.1142/9789813272880_0206}}
 
 @article{albi_vehicular_2019,
 	abstract = {This paper presents a review and critical analysis on the modeling of the dynamics of vehicular traffic, human crowds and swarms seen as living and, hence, complex systems. It contains a survey of the kinetic models developed in the last 10 years on the aforementioned topics so that overlapping with previous reviews can be avoided. Although the main focus of this paper lies on the mesoscopic models for collective dynamics, we provide a brief overview on the corresponding micro and macroscopic models, and discuss intermediate role of mesoscopic model between them. Moreover, we provide a number of selected challenging research perspectives for readers' attention.},
@@ -234,7 +239,8 @@
 	title = {Alignment of {Self}-propelled {Rigid} {Bodies}: {From} {Particle} {Systems} to {Macroscopic} {Equations}},
 	url = {http://link.springer.com/10.1007/978-3-030-15096-9_2},
 	year = {2019},
-	Bdsk-Url-1 = {http://link.springer.com/10.1007/978-3-030-15096-9_2}}
+	Bdsk-Url-1 = {http://link.springer.com/10.1007/978-3-030-15096-9_2},
+	doi = {10.1007/978-3-030-15096-9_2}}
 
 @article{degond_nematic_2020,
[…]
