### 2021-08-23 #3425 @pdebuyl
@junghans would you review the article "SiSyPHE: A Python package for the Simulation of Systems of Particles with High Efficiency" for the Journal of Open Source Software ?
