### 2021-08-24 #3653 @whedon
[REVIEW]: SiSyPHE: A Python package for the Simulation of Systems of interacting mean-field Particles with High Efficiency
**Submitting author:** @antoinediez (<a href="http://orcid.org/0000-0002-4454-8452">Antoine Diez</a>)
**Repository:** <a href="https://github.com/antoinediez/Sisyphe" target ="_blank">https://github.com/antoinediez/Sisyphe</a>
**Version:** v1.0
