### 2021-08-24 8b6f0eb
add missing DOIs
