+++ doc/biblio_sisyphe.bib
@@ -54,7 +54,8 @@
 	pages = {165--251},
 	publisher = {Springer},
 	title = {Topics in propagation of chaos},
-	year = {1991}}
+	year = {1991}
+	doi = {10.1007/bfb0085169}}
 
 @incollection{meleard_asymptotic_1996,
 	author = {M{\'e}l{\'e}ard, S.},
@@ -368,7 +369,8 @@
 	title = {From particle swarm optimization to consensus based optimization: stochastic modeling and mean-field limit},
 	url = {http://arxiv.org/abs/2012.05613},
 	year = {2020},
-	Bdsk-Url-1 = {http://arxiv.org/abs/2012.05613}}
+	Bdsk-Url-1 = {http://arxiv.org/abs/2012.05613}
+	doi = {10.1142/s0218202521500342}}
 
 @article{mei_mean_2018,
 	abstract = {Multilayer neural networks are among the most powerful models in machine learning, yet the fundamental reasons for this success defy mathematical understanding. Learning a neural network requires optimizing a nonconvex high-dimensional objective (risk function), a problem that is usually attacked using stochastic gradient descent (SGD). Does SGD converge to a global optimum of the risk or only to a local optimum? In the former case, does this happen because local minima are absent or because SGD somehow avoids them? In the latter, why do local minima reached by SGD have good generalization properties? In this paper, we consider a simple case, namely two-layer neural networks, and prove that---in a suitable scaling limit---SGD dynamics is captured by a certain nonlinear partial differential equation (PDE) that we call distributional dynamics (DD). We then consider several specific examples and show how DD can be used to prove convergence of SGD to networks with nearly ideal generalization error. This description allows for ``averaging out'' some of the complexities of the landscape of neural networks and can be used to prove a general convergence result for noisy SGD.},
