### 2021-08-24 2648ab3
add missing DOIs
