+++ doc/biblio_sisyphe.bib
@@ -54,7 +54,7 @@
 	pages = {165--251},
 	publisher = {Springer},
 	title = {Topics in propagation of chaos},
-	year = {1991}
+	year = {1991},
 	doi = {10.1007/bfb0085169}}
 
 @incollection{meleard_asymptotic_1996,
@@ -369,7 +369,7 @@
 	title = {From particle swarm optimization to consensus based optimization: stochastic modeling and mean-field limit},
 	url = {http://arxiv.org/abs/2012.05613},
 	year = {2020},
-	Bdsk-Url-1 = {http://arxiv.org/abs/2012.05613}
+	Bdsk-Url-1 = {http://arxiv.org/abs/2012.05613},
 	doi = {10.1142/s0218202521500342}}
 
 @article{mei_mean_2018,
