### 2021-08-30 antoinediez/Sisyphe#1 @junghans
Add contribution guidelines
For https://github.com/openjournals/joss-reviews/issues/3653 please a `CONTRIBUTING` will to your repo.
(see https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors for technical details.)
