### 2021-08-30 antoinediez/Sisyphe#2 @junghans
Add CI / improve testing
Related to https://github.com/openjournals/joss-reviews/issues/3653

The README only refers to one test `sisyphe.test_sisyphe()`, which takes a long time, only tests some functionality and needs to be run manually.

It would be good to add CI, e.g. GitHub action to repo and run these tests at every push.

Testing python packages is pretty simple,  see https://docs.github.com/en/actions/guides/building-and-testing-python
