### 2021-08-30 antoinediez/Sisyphe#3 @junghans
JOSS Paper title to broad
Related to https://github.com/openjournals/joss-reviews/issues/3653.

Coming from the molecular dynamics (MD) background the title of the JOSS Paper _A Python package for the Simulation of Systems of Particles with High Efficiency_ seems a bit too broad as MD is even mentioned at all, maybe a more toned-down title like _A Python package for the Simulation of Systems of interacting mean-field Particles with High Efficiency_ would be better.
