### 2021-08-31 antoinediez/Sisyphe#3 @pdebuyl
@antoinediez keep in mind that the title of the paper and of the zenodo archive of the repository must be identical.
