+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# How to contribute
+
+Thank you for your interest in SiSyPHE, contributions to make the library grow are warmly welcome! 
+
+## Examples of contributions
+
+Examples of possible (and ongoing) developments to which you are welcome to contribute include the following. 
+
+* The implementation of new models.
+
+* The implementation of more complex boundary conditions and of models on *non-flat* manifolds. 
+
+* An improved visualization method (currently only basic visualization functions relying on [Matplotlib](https://matplotlib.org/) are implemented). 
+
+## Submitting changes 
+
+If you find a bug or an unexpected behaviour, please open an issue on Github. To add a new feature to the library, please send a pull request with a clear description of what you have done. If you are working on a model which you would like to be included in SiSyPHE, I will be happy to discuss it beforehand either directly on Github or by email first. 
