### 2021-08-31 antoinediez/Sisyphe#1 @antoinediez
A contributing guidelines file has been added.
