### 2021-08-31 antoinediez/Sisyphe#3 @antoinediez
Thank you for your feedback!
I do agree, the title that you propose is more precise.
@pdebuyl  I changed the title of the paper and I will make sure that the title of the Zenodo archive will be the same. However I don't think that I can change the title of the JOSS issue, but I don't know if it is necessary?
