### 2021-08-31 antoinediez/Sisyphe#3 @junghans
Updated the issue!

Where is Zenodo pulling the title from? Is it from the repository description? (Which is still "Simulation of Systems of Particles with High Efficiency")
