### 2021-09-02 antoinediez/Sisyphe#3 @antoinediez
I updated the repository description.

Once the paper will (hopefully) be accepted, I will make a new release and the title of the Zenodo archive will be the title of this release. So it shouldn't be a problem.
