### 2021-09-02 antoinediez/Sisyphe#2 @antoinediez
Thanks for your feedback.

The function `sisyphe.test_sisyphe()` is designed to test the core functionalities of the package on a "real-world example". As a consequence, it is indeed computationally demanding, but in the other hand, I chose a model which is theoretically well understood so that it is possible to check the result of the test with accuracy.

In the coming days, I will nevertheless add CI to the package to test more functionalities on smaller examples.  Unfortunately, the Github hosted machines do not have a GPU so the tests will be limited to much simpler, small scale cases.
