### 2021-09-02 antoinediez/Sisyphe#2 @junghans
Yeah, I ran `sisyphe.test_sisyphe()` on my laptop (without GPU) and it took a long time there, too. Not great for a quick test if the package works ;-)
