+++ requirements.txt
@@ -1,0 +1,2 @@
+ # installs dependencies from ./setup.py, and the package itself
+.
