### 2021-09-04 2920351
add quick testing functions
