+++ sisyphe/__init__.py
@@ -1,3 +1,6 @@
 import os
 
 from .test.test_script import test_sisyphe
+from .test.quick_test import test_neighbours
+from .test.quick_test import test_bsr
+from .test.quick_test import test_simu
