+++ .github/workflows/testing.yaml
@@ -30,5 +30,5 @@ jobs:
         pip install -r requirements.txt
     - name: Test with pytest
       run: |
-	pip install pytest
+        pip install pytest
         pytest sisyphe/test/quick_test.py
