+++ .github/workflows/testing.yaml
@@ -8,7 +8,8 @@ on:
     branches: [ $default-branch ]
   pull_request:
     branches: [ $default-branch ]
-
+  workflow_dispatch:
+    branches: [ $default-branch ]
 jobs:
   build:
 
