### 2021-09-04 99e39f9
add matplotlib in the dependencies
