+++ setup.py
@@ -51,6 +51,7 @@ setup(
     install_requires=[
         "numpy",
         "scipy",
+        "matplotlib",
         "torch",
         "pykeops",
     ],
