+++ setup.py
@@ -15,7 +15,7 @@ with open(path.join(here, "README.md"), encoding="utf-8") as f:
 
 setup(
     name="sisyphe",
-    version="0.1.5",
+    version="0.1.6",
     description="Simulation of Systems of interacting mean-field Particles with High Efficiency",  # Required
     long_description=long_description,
     long_description_content_type="text/markdown",
