### 2021-09-04 a4def58
describe the quick testing procedure in the doc
