### 2021-09-05 1789b0b
add more quick tests
