+++ sisyphe/__init__.py
@@ -3,4 +3,7 @@ import os
 from .test.test_script import test_sisyphe
 from .test.quick_test import test_neighbours
 from .test.quick_test import test_bsr
-from .test.quick_test import test_simu
+from .test.quick_test import test_vicsek
+from .test.quick_test import test_dorsogna
+from .test.quick_test import test_volume_exclusion
+
