### 2021-09-05 d67c343
fix depreciation warning with factorial
