### 2021-09-05 antoinediez/Sisyphe#2 @antoinediez
I added the quick test module `sisyphe.test.quick_test`. The various test functions check that the core functionalities work and run small scale simulations (taken form the example gallery) to ensure that the code runs without error. These tests are run automatically at every push as suggested by @junghans. I also updated the installation page of the documentation to warn the user that the more advanced test function `sisyphe.test_sisyphe()` may take a long time.

Please tell me if the addition of this module answers your comment or if you would like to see something else. I won't close this issue until both of you (@junghans and @lorenzo-rovigatti) tick the "Automated tests" item on the review page.
