### 2021-09-13 #3653 @pdebuyl
PS: @lorenzo-rovigatti check the issue there as well https://github.com/antoinediez/Sisyphe/issues/2
