### 2021-09-15 #3653 @pdebuyl
Hi @lorenzo-rovigatti I want to make sure that you can proceed. JOSS recommends less than 6 weeks for the review, so there is no timing issue. Thanks for the update though.
