### 2021-09-18 antoinediez/Sisyphe#2 @lorenzo-rovigatti
Thanks to @junghans for opening this issue and @antoinediez for providing a solution. Unfortunately I'm having some issues with it. I have installed sisyphe by cloning the repo and using `pip3 install .` in the repo's folder. I'm using Ubuntu, Python 3.8.10, pytest 6.2.5.

The following snippet

```python
import pytest
from sisyphe.test import quick_test
retcode = pytest.main(quick_test.__file__)
```

results in

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/lorenzo/.local/lib/python3.8/site-packages/_pytest/config/__init__.py", line 143, in main
    config = _prepareconfig(args, plugins)
  File "/home/lorenzo/.local/lib/python3.8/site-packages/_pytest/config/__init__.py", line 307, in _prepareconfig
    raise TypeError(msg.format(args, type(args)))
TypeError: `args` parameter expected to be a list of strings, got: '/home/lorenzo/SCRATCH/JOSS_REVIEWS/Sisyphe/sisyphe/test/quick_test.py' (type: <class 'str'>)
```

I never used Pytest but looking at the error I tried to fix it with `retcode = pytest.main([quick_test.__file__, ])` and the test procedure starts.

Unfortunately this results in quite a few `FileExistsError` errors. They seem trivial but I cannot get around them. I attach the output to this post.

[test_output.txt](https://github.com/antoinediez/Sisyphe/files/7189877/test_output.txt)
