### 2021-09-18 antoinediez/Sisyphe#2 @antoinediez
Hi @lorenzo-rovigatti !

Thanks for correcting `retcode = pytest.main(quick_test.__file__, )` into `retcode = pytest.main([quick_test.__file__, ])`. The previous syntax is depreciated and does not work with the latest version of Pytest. I will update the documentation.

For the second (and more critical) issue, I think that it comes from PyKeOps. What does the following snippet output on your machine ?
```python
import pykeops
pykeops.clean_pykeops()
pykeops.test_torch_bindings()
```
If it does not work then, I think that it is related to this issue https://github.com/getkeops/keops/issues/142#. This should be solved by upgrading Cmake to Cmake>=3.18.

To check that, I tried the following: on a Google Colab session, typing
```python
!pip install sisyphe
import pytest
from sisyphe.test import quick_test
retcode = pytest.main([quick_test.__file__,])
```
gives exactly the same error as yours. However, by typing first
```python
!pip install pykeops[colab]
```
this will upgrade Cmake to Cmake 3.18 before installing PyKeOps and the tests then run as expected.
