### 2021-09-18 cdf01d9
fix syntax to manually run the quick tests
