### 2021-09-18 antoinediez/Sisyphe#2 @lorenzo-rovigatti
Yes, the snippet does not work. I'll try to update cmake and see if that solves the issue!
