### 2021-09-18 antoinediez/Sisyphe#2 @lorenzo-rovigatti
Updating cmake does solve the issue. Maybe you want to add this requirement somewhere in the documentation. Either than that I think you mark this issue as resolved!
