### 2021-09-18 antoinediez/Sisyphe#2 @antoinediez
Great thank you!
I did not add Cmake in the requirements because it is already a requirement of PyKeOps. It is likely that the authors will add Cmake>=3.18 as a requirement in a future release (or find another way to solve the issue). In the meantime, I will add a note in the documentation to warn the user that upgrading Cmake may be necessary.
