### 2021-09-18 8699643
add a note on the required cmake version
