### 2021-09-20 #3653 @pdebuyl
Hi @antoinediez I will now proceed to the editorial steps outlined here: https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance
