### 2021-09-20 #3653 @lorenzo-rovigatti
@pdebuyl sure. I'm happy to recommend acceptance of SiSyPHE in JOSS.
