### 2021-09-27 #3653 @pdebuyl
@antoinediez thank you for the submission. The package is very nice and well documented. I have a few minor comments on the article:

1. The "statement of need" section is a bit mixed-up with the presentation of the package.
2. In the same spirit: the summary does not state explicitly what SiSyPHE is. Can you edit the text so that the summary is a summary and likewise for the statement of need?
3. The focus on biologically-based simulations neglects the field of Molecular Dynamics in physics and chemistry. The sentence "One of the main motivations is the modelling of large animal societies and the emergence10of complex patterns from simple behavioral rules" leads the reader to think that particle-based simulations are only relevant for biology. You don't specifically need to give examples of physics or chemistry simulations, but make it clear that the "main motivation" that you give is within the biological sciences.
4. Can you mention briefly the type of time-stepping algorithms that are used in the package? This information is useful to give readers a good overview of what's in the package.
5. The book by Del Moral has a ISBN, please include it https://www.routledge.com/Mean-Field-Simulation-for-Monte-Carlo-Integration/Del-Moral/p/book/9781466504059 (if you refer to the ebook, there is also a ISBN or a DOI).
6. Grassi & Pareschi has a journal reference, please use it.

You will need to deposit the package on zenodo before the publication, let me know if you need assistance about this.
