### 2021-09-27 bb355e8
add missing journal and isbn
