+++ doc/biblio_sisyphe.bib
@@ -2,7 +2,7 @@
 %% http://bibdesk.sourceforge.net/
 
 
-%% Created for Antoine Diez at 2021-06-27 22:29:00 +0200 
+%% Created for Antoine Diez at 2021-09-27 14:57:59 +0200 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
@@ -43,23 +43,25 @@
 	url = {https://linkinghub.elsevier.com/retrieve/pii/0021999187901409},
 	volume = {73},
 	year = {1987},
-	Bdsk-Url-1 = {https://linkinghub.elsevier.com/retrieve/pii/0021999187901409},
-	Bdsk-Url-2 = {https://doi.org/10.1016/0021-9991(87)90140-9}}
+	bdsk-url-1 = {https://linkinghub.elsevier.com/retrieve/pii/0021999187901409},
+	bdsk-url-2 = {https://doi.org/10.1016/0021-9991(87)90140-9}}
 
 @incollection{sznitman_topics_1991,
 	author = {Sznitman, A.-S.},
 	booktitle = {{\'E}c. {\'E}t{\'e} {Probab}. {St}.-{Flour} {XIX}---1989},
+	doi = {10.1007/bfb0085169},
 	file = {Sznitman - 1991 - Topics in propagation of chaos.pdf:/Users/AntoineDiez/Zotero/storage/J5INTEZ8/Sznitman - 1991 - Topics in propagation of chaos.pdf:application/pdf},
 	keywords = {McKean-Vlasov, Lipschitz, Convolution},
 	pages = {165--251},
 	publisher = {Springer},
 	title = {Topics in propagation of chaos},
 	year = {1991},
-	doi = {10.1007/bfb0085169}}
+	bdsk-url-1 = {https://doi.org/10.1007/bfb0085169}}
 
 @incollection{meleard_asymptotic_1996,
 	author = {M{\'e}l{\'e}ard, S.},
 	booktitle = {Probabilistic {Models} for {Nonlinear} {Partial} {Differential} {Equations}},
+	doi = {10.1007/bfb0093177},
 	editor = {Talay, D. and Tubaro, L.},
 	file = {M{\'e}l{\'e}ard - 1996 - Asymptotic Behaviour of some interacting particle .pdf:/Users/AntoineDiez/Zotero/storage/YKTTN9AV/M{\'e}l{\'e}ard - 1996 - Asymptotic Behaviour of some interacting particle .pdf:application/pdf},
 	keywords = {Boltzmann, Jumps, McKean-Vlasov, Lipschitz, Convolution},
@@ -68,7 +70,7 @@
 	series = {Lecture {Notes} in {Mathematics}},
 	title = {Asymptotic {Behaviour} of some interacting particle systems; {McKean}-{Vlasov} and {Boltzmann} models},
 	year = {1996},
-	doi = {10.1007/bfb0093177}}
+	bdsk-url-1 = {https://doi.org/10.1007/bfb0093177}}
 
 @article{vicsek_novel_1995,
 	author = {Vicsek, T. and Czir{\'o}k, A. and Ben-Jacob, E. and Cohen, I. and Shochet, O.},
@@ -81,8 +83,8 @@
 	url = {https://link.aps.org/doi/10.1103/PhysRevLett.75.1226},
 	volume = {75},
 	year = {1995},
-	Bdsk-Url-1 = {https://link.aps.org/doi/10.1103/PhysRevLett.75.1226},
-	Bdsk-Url-2 = {https://doi.org/10.1103/PhysRevLett.75.1226}}
+	bdsk-url-1 = {https://link.aps.org/doi/10.1103/PhysRevLett.75.1226},
[…]
