+++ doc/biblio_sisyphe.bib
@@ -2,12 +2,44 @@
 %% http://bibdesk.sourceforge.net/
 
 
-%% Created for Antoine Diez at 2021-09-27 14:57:59 +0200 
+%% Created for Antoine Diez at 2021-09-27 16:13:00 +0200 
 
 
 %% Saved with string encoding Unicode (UTF-8) 
 
+@article{piggott,
+	author = {Piggott, M. J. and Solo, V.},
+	doi = {10.1137/15M1019726},
+	journal = {SIAM J. Numer. Anal.},
+	number = {4},
+	pages = {2490--2516},
+	title = {Geometric {E}uler--{M}aruyama Schemes for Stochastic Differential Equations in {SO(n)} and {SE(n)}},
+	volume = {54},
+	year = {2016}}
 
+@book{kloeden,
+	author = {Kloeden, Peter E. and Platen, Eckhard},
+	date-added = {2021-09-27 16:10:29 +0200},
+	date-modified = {2021-09-27 16:12:48 +0200},
+	publisher = {Springer, Berlin, Heidelberg},
+	series = {Applications of Mathematics},
+	number = {23},
+	title = {Numerical Solution of Stochastic Differential Equations},
+	year = {1992},
+	isbn = {978-3-540-54062-5},
+	doi = {10.1007/978-3-662-12616-5}}
+
+@book{leihmkuhler,
+	author = {Leimkuhler, Ben and Matthews, Charles },
+	date-added = {2021-09-27 16:06:28 +0200},
+	date-modified = {2021-09-27 16:08:32 +0200},
+	isbn = {978-3-319-16374-1},
+	number = {39},
+	publisher = {Springer International Publishing},
+	series = {Interdisciplinary Applied Mathematics},
+	title = {Molecular Dynamics With Deterministic and Stochastic Numerical Methods},
+	year = {2015},
+	doi = {10.1007/978-3-319-16375-8}}
 
 @inproceedings{kac_foundations_1956,
 	author = {Kac, Mark},
@@ -143,12 +175,12 @@
 @book{del_moral_mean_2013,
 	author = {Del Moral, P.},
 	file = {Del Moral - 2016 - Mean field simulation for Monte Carlo integration.pdf:/Users/AntoineDiez/Zotero/storage/XRJNBV7J/Del Moral - 2016 - Mean field simulation for Monte Carlo integration.pdf:application/pdf},
+	isbn = {9781466504059},
 	number = {126},
 	publisher = {CRC Press, Taylor \& Francis Group},
 	series = {Monographs on {Statistics} and {Applied} {Probability}},
 	title = {Mean field simulation for {Monte} {Carlo} integration},
-	year = {2013},
-	isbn = {9781466504059}}
+	year = {2013}}
 
 @article{hauray_kacs_2014,
 	author = {Hauray, M. and Mischler, S.},
@@ -573,6 +605,7 @@
 	doi = {10.1007/978-1-4757-3437-9},
 	editor = {Doucet, Arnaud and Freitas, Nando and Gordon, Neil},
 	file = {Doucet et al. - 2001 - Sequential Monte Carlo Methods in Practice.pdf:/Users/AntoineDiez/Zotero/storage/4C2QU7X6/Doucet et al. - 2001 - Sequential Monte Carlo Methods in Practice.pdf:application/pdf},
+	isbn = {978-0-387-95146-1},
 	publisher = {Springer-Verlag New York},
 	series = {Information {Science} and {Statistics}},
 	title = {Sequential {Monte} {Carlo} {Methods} in {Practice}},
@@ -580,8 +613,7 @@
 	urldate = {2021-05-13},
[…]
