+++ paper.md
@@ -22,7 +22,7 @@ bibliography: doc/biblio_sisyphe.bib
 
 Over the past decades, the study of systems of particles has become an important 
 part of many research areas, from theoretical physics to applied biology and 
-computational mathematics. One of the main motivations is the modelling of large
+computational mathematics. One of the main motivations in Mathematical Biology is the modelling of large
 animal societies and the emergence of complex patterns from simple behavioral rules, e.g. the flock of birds, fish schools, ant colonies etc. In the microscopic
 world, particle systems are used to model a wide range of phenomena, 
 from the collective motion of spermatozoa to the anarchical development of 
@@ -34,20 +34,27 @@ These models are usually based on Partial Differential Equations (PDE) derived
 using phenomenological considerations which are often difficult to justify 
 mathematically [@degond_continuum_2008; @dimarco_self-alignment_2016; @degond_bulk_2021]. Finally, inspired by models in Biology, there is an ever growing
 literature on the design of algorithms based on the simulation of *artificial*
-particle systems to solve tough optimization problems [@kennedy_particle_1995; @pinnau_consensus-based_2017; @totzeck_trends_2021; @grassi_particle_2020] and to construct new more efficient Markov Chain Monte Carlo methods [@del_moral_measure-valued_1998; @del_moral_mean_2013; @doucet_sequential_2001; @cappe_population_2004; @clarte_collective_2021].
+particle systems to solve tough optimization problems [@kennedy_particle_1995; @pinnau_consensus-based_2017; @totzeck_trends_2021; @grassi_particle_2020] and to construct new more efficient Markov Chain Monte Carlo methods [@del_moral_measure-valued_1998; @del_moral_mean_2013; @doucet_sequential_2001; @cappe_population_2004; @clarte_collective_2021]. The simulation of systems of particles is also at the core of Molecular Dynamics [@leihmkuhler], although the present library is not specifically written for this purpose. The `SiSyPHE` library builds on recent advances in hardware and software 
+for the efficient simulation of large scale interacting *mean-field* particle systems, 
+both on the GPU and on the CPU. The versatile object-oriented Python interface of the library is designed for the simulation and comparison of new and classical many-particle models of collective dynamics in Mathematics and Active Matter Physics, enabling ambitious numerical experiments and leading to novel conjectures and results.
 
 # Statement of need
 
-The `SiSyPHE` library builds on recent advances in hardware and software 
-for the efficient simulation of large scale interacting *mean-field* particle systems, 
-both on the GPU and on the CPU. The implementation is based on recent libraries 
+A major difficulty in the simulation of systems of particles is the high computational cost, typically quadratic in the number of particles, which prevents large scale experiments. The implementation of `SiSyPHE` is based on recent libraries 
 originally developed for machine learning purposes to significantly accelerate 
-tensor (array) computations, namely the `PyTorch` package [@paszke_pytorch_2019] and the `KeOps` library [@charlier_kernel_2021]. 
-The versatile object-oriented Python interface is well suited to the comparison 
-of new and classical many-particle models, enabling ambitious numerical 
-experiments and leading to novel conjectures. The SiSyPHE library speeds up 
-both traditional Python and low-level implementations by one to three orders 
+tensor (array) computations, namely the `PyTorch` package [@paszke_pytorch_2019] and the `KeOps` library [@charlier_kernel_2021]. On a GPU, the `SiSyPHE` library speeds up both traditional Python and low-level implementations by one to three orders 
 of magnitude for systems with up to several millions of particles. 
+
[…]
