### 2021-09-27 #3653 @antoinediez
Thank you all!

@pdebuyl I updated the paper accordingly to your comments, please tell me if you find it better now or if you think it needs other modifications. Once you will approve the paper, I will then deposit the package on Zenodo.
