### 2021-09-28 f6fa67e
minor updates to paper
