+++ paper.md
@@ -22,21 +22,21 @@ bibliography: doc/biblio_sisyphe.bib
 
 Over the past decades, the study of systems of particles has become an important 
 part of many research areas, from theoretical physics to applied biology and 
-computational mathematics. One of the main motivations in Mathematical Biology is the modelling of large
-animal societies and the emergence of complex patterns from simple behavioral rules, e.g. the flock of birds, fish schools, ant colonies etc. In the microscopic
+computational mathematics. One of the main motivations in mathematical biology is the modelling of large
+animal societies and the emergence of complex patterns from simple behavioral rules, e.g., flocks of birds, fish schools, ant colonies, etc. In the microscopic
 world, particle systems are used to model a wide range of phenomena, 
 from the collective motion of spermatozoa to the anarchical development of 
 cancer cells. Within this perspective, there are at least three important reasons to conduct large scale computer simulations of particle systems. First, numerical experiments are essential to 
 calibrate the models and test the influence of each parameter in a controlled 
 environment. For instance, the renowned Vicsek model [@vicsek_novel_1995]
-is a minimal model of *flocking* which exhibits a complex behavior, studied numerically in particular in [@chate_collective_2008]. Secondly, particle simulations are used to check the validity of *macroscopic* models which describe the statistical behavior of particle systems.
-These models are usually based on Partial Differential Equations (PDE) derived 
-using phenomenological considerations which are often difficult to justify 
-mathematically [@degond_continuum_2008; @dimarco_self-alignment_2016; @degond_bulk_2021]. Finally, inspired by models in Biology, there is an ever growing
+is a minimal model of *flocking*, which exhibits a complex behavior, studied numerically in particular in [@chate_collective_2008]. Secondly, particle simulations are used to check the validity of *macroscopic* models that describe the statistical behavior of particle systems.
+These models are usually based on partial differential equations (PDE) derived 
+using phenomenological considerations that are often difficult to justify 
+mathematically [@degond_continuum_2008; @dimarco_self-alignment_2016; @degond_bulk_2021]. Finally, inspired by models in biology, there is an ever growing
 literature on the design of algorithms based on the simulation of *artificial*
-particle systems to solve tough optimization problems [@kennedy_particle_1995; @pinnau_consensus-based_2017; @totzeck_trends_2021; @grassi_particle_2020] and to construct new more efficient Markov Chain Monte Carlo methods [@del_moral_measure-valued_1998; @del_moral_mean_2013; @doucet_sequential_2001; @cappe_population_2004; @clarte_collective_2021]. The simulation of systems of particles is also at the core of Molecular Dynamics [@leihmkuhler], although the present library is not specifically written for this purpose. The `SiSyPHE` library builds on recent advances in hardware and software 
+particle systems to solve tough optimization problems [@kennedy_particle_1995; @pinnau_consensus-based_2017; @totzeck_trends_2021; @grassi_particle_2020] and to construct new more efficient Markov Chain Monte Carlo methods [@del_moral_measure-valued_1998; @del_moral_mean_2013; @doucet_sequential_2001; @cappe_population_2004; @clarte_collective_2021]. The simulation of systems of particles is also at the core of molecular dynamics [@leihmkuhler], although the present library is not specifically written for this purpose. The `SiSyPHE` library builds on recent advances in hardware and software 
 for the efficient simulation of large scale interacting *mean-field* particle systems, 
[…]
