### 2021-09-28 #3653 @danielskatz
@antoinediez - I've suggested some minor changes to the paper in https://github.com/antoinediez/Sisyphe/pull/5 - please merge this, or let me know what you disagree with, then we can proceed to acceptance and publication.
